Added detailed messages to settings cloning
Added ability to eject unpackager trackers
Added ability to clear recipe holders and markers in crafting grid
Added settings cloning to fluid package filler
Changed saving settings to cloner to use shift-right click

+++ src/main/java/thelm/packagedauto/client/screen/VolumeAmountSpecifyingScreen.java
@@ -45,7 +45,7 @@ public class VolumeAmountSpecifyingScreen extends AmountSpecifyingScreen<VolumeA
 			int amount = Mth.clamp(Integer.parseInt(amountField.getValue()), 0, maxAmount);
 			IVolumeStackWrapper newStack = stack.copy();
 			newStack.setAmount(amount);
-			PacketHandler.INSTANCE.sendToServer(new SetItemStackPacket((short)containerSlot, MiscHelper.INSTANCE.makeVolumePackage(newStack)));
+			PacketHandler.INSTANCE.sendToServer(new SetItemStackPacket(containerSlot, MiscHelper.INSTANCE.makeVolumePackage(newStack)));
 			close();
 		}
 		catch(NumberFormatException e) {
